# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: نظام إدارة توزيع السلع - Distribution Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /distributors:
    get:
      operationId: getDistributors
      tags: [distributors]
      summary: Get all distributors
      responses:
        "200":
          description: List of distributors
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Distributor"
    post:
      operationId: createDistributor
      tags: [distributors]
      summary: Create distributor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDistributorRequest"
      responses:
        "201":
          description: Distributor created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
  /distributors/{id}:
    get:
      operationId: getDistributor
      tags: [distributors]
      summary: Get distributor by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Distributor details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
    put:
      operationId: updateDistributor
      tags: [distributors]
      summary: Update distributor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDistributorRequest"
      responses:
        "200":
          description: Updated distributor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Distributor"
    delete:
      operationId: deleteDistributor
      tags: [distributors]
      summary: Delete distributor
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /distributors/{id}/location:
    put:
      operationId: updateDistributorLocation
      tags: [distributors]
      summary: Update distributor location (GPS tracking)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LocationUpdate"
      responses:
        "200":
          description: Location updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: Get all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /accounting/summary:
    get:
      operationId: getAccountingSummary
      tags: [accounting]
      summary: Get accounting summary (profits, debts)
      parameters:
        - name: period
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Accounting summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountingSummary"
...
  /accounting/settle/{distributorId}:
    post:
      operationId: settleDistributorAccount
      tags: [accounting]
      summary: Settle distributor account
      parameters:
        - name: distributorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettleAccountRequest"
      responses:
        "200":
          description: Account settled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    Distributor:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        username:
          type: string
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        lastSeen:
          type: string
          format: date-time
          nullable: true
        totalTasksCompleted:
          type: integer
        totalAmountCollected:
          type: number
        debt:
          type: number
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - phone
        - isActive
        - username
        - totalTasksCompleted
        - totalAmountCollected
        - debt
        - createdAt
    CreateDistributorRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        username:
          type: string
        password:
          type: string
      required:
        - firstName
        - lastName
        - phone
        - username
        - password
    UpdateDistributorRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        isActive:
          type: boolean
        password:
          type: string
    LocationUpdate:
      type: object
      properties:
        latitude:
          type: number
        longitude:
          type: number
      required:
        - latitude
        - longitude
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        purchasePrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
        isLowStock:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - purchasePrice
        - sellPrice
        - quantity
        - lowStockThreshold
        - isLowStock
        - createdAt
    CreateProductRequest:
      type: object
      properties:
        name:
          type: string
        purchasePrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
      required:
        - name
        - purchasePrice
        - sellPrice
        - quantity
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        purchasePrice:
          type: number
        sellPrice:
          type: number
        quantity:
          type: integer
        lowStockThreshold:
          type: integer
...
    AccountingSummary:
      type: object
      properties:
        period:
          type: string
        totalRevenue:
          type: number
        totalCost:
          type: number
        totalProfit:
          type: number
        totalStoreDebts:
          type: number
        totalDistributorDebts:
          type: number
        totalDeliveries:
          type: integer
        totalTasks:
          type: integer
      required:
        - period
        - totalRevenue
        - totalCost
        - totalProfit
        - totalStoreDebts
        - totalDistributorDebts
        - totalDeliveries
        - totalTasks
...
    SettleAccountRequest:
      type: object
      properties:
        amount:
          type: number
        notes:
          type: string
      required:
        - amount
...